Added a main menu and some improvements to the level
Added a main menu of sorts to choose which level to load. For now new levels can be added to the script handling the level loading until I have found a better way of handling the levels. Also will make the buttons and the layout better as well in the future :)